Refactor: Extracting guides_helper
