+++ app/helpers/application_helper.rb
@@ -34,18 +34,6 @@ module ApplicationHelper
     'class="active"'.html_safe if expected == @current_tab
   end
 
-  def stats(stats, k)
-    "#{stats.send k} #{status_icon(k)} ".html_safe
-  end
-
-  def practice_key_for(stats)
-    if stats.started?
-      :continue_practicing
-    else
-      :start_practicing
-    end
-  end
-
   def time_ago_in_words_or_never(date)
     date ? time_ago_in_words(date) : t(:never)
   end
