--- app/helpers/user_guides_helper.rb
@@ -1,2 +1,0 @@
-module UserGuidesHelper
-end
